Nextauth test
