+++ frontend/.env.local
@@ -1,0 +1,4 @@
+AZURE_AD_CLIENT_ID=7e685700-a59e-4f58-9d49-7db2d7e8fc63
+AZURE_AD_CLIENT_SECRET=509469ba-b1aa-4309-a372-4e5ac7863b82
+AZURE_AD_TENANT_ID=7a3c7284-54fb-45b7-9bdc-4ed61f8ed7bc
+NEXTAUTH_URL=http://localhost:3000
