+++ frontend/src/app/api/auth/[...nextauth]/route.js
@@ -1,0 +1,17 @@
+import NextAuth from "next-auth"
+import AzureADProvider from "next-auth/providers/azure-ad"
+
+export const authOptions = {
+  providers: [
+    AzureADProvider({
+      clientId: process.env.AZURE_AD_CLIENT_ID,
+      clientSecret: process.env.AZURE_AD_CLIENT_SECRET,
+      tenantId: process.env.AZURE_AD_TENANT_ID,
+    }),
+  ],
+}
+
+// App Router 用：HTTP メソッドごとの named export
+const handler = NextAuth(authOptions)
+
+export { handler as GET, handler as POST }
