+++ frontend/src/components/SignInButton.tsx
@@ -1,0 +1,23 @@
+'use client'
+
+import { useSession, signIn, signOut } from 'next-auth/react'
+
+export default function SignInButton() {
+  const { data: session } = useSession()
+
+  if (session) {
+    return (
+      <>
+        <p>Signed in as {session.user?.email}</p>
+        <button onClick={() => signOut()}>Sign out</button>
+      </>
+    )
+  }
+
+  return (
+    <>
+      <p>Not signed in</p>
+      <button onClick={() => signIn('azure-ad')}>Sign in with Microsoft</button>
+    </>
+  )
+}
